copy_dashboard function variable dup_dashboard_id
This error happens when res returns false, with a 503 error, when using an Heroku instance. 
It will depend on the number and size of Heroku dynos. In my case this issue happened if the Dashboard contained more than 24 cards/questions.
However the dashboard was created. We just need to pick the new dashboard id.